feat: Provide a GUI to set up the parameter for generating the documentation and to control the flow of the process.
Add a new module calles 'gui.py' in the main root folder to define the layout and behaviour of the GUI (windows) which enables the user to control the whole process and set up the different parameters. The relevant class in this modul is designed in such a way that it can be referenced by the main modul, but it doesn't use any data from other parts of the existing program (main). So the process is controlled completly by the main function of the main module. Furthermore this changes made some more changes in the design of the existing code necessary: For loading the parameters from the gui class into the MetaDataClass in the main module some parts of the code has been modified to  get the desired behaviour: The MetaData class will not be initialized automatically directly after implemention anymore by inharencing from AutoCall-class. Instead the method is called explicitly.