refactor:#6 Implement a method to find multiline-comments and store their line numbers. (not accessed yet)
Therefore refactor initialize-method to read the input-source-text as str in the first place. To use the current structure this str is splittet at line ends. But additional the str is used in the new method to search for multiline comments (by now only for python syntax), and store all line-numbers (0-based-index) in a list. Further TODO would be to query wheather a found reference or calling-sequence is in this line-number, if yes it should be ignored.